split statewide testing analysis into separate iteration
